BekoSIRS_api/requirements.txt
# BekoSIRS Backend - Python Dependencies
# Pin all versions for reproducible builds

# ==============================================================================
# CORE FRAMEWORK
# ==============================================================================
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# ==============================================================================
# API DOCUMENTATION
# ==============================================================================
drf-spectacular==0.29.0

# ==============================================================================
# CORS & SECURITY
# ==============================================================================
django-cors-headers==4.3.1

# ==============================================================================
# ENVIRONMENT & CONFIGURATION
# ==============================================================================
python-dotenv==1.0.0

# ==============================================================================
# DATABASE DRIVERS
# ==============================================================================
# For production with Microsoft SQL Server
mssql-django==1.3
# pyodbc is required for mssql-django (install system ODBC driver separately)
# pyodbc==5.0.1

# ==============================================================================
# ML & DATA PROCESSING
# ==============================================================================
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1

# ==============================================================================
# FILE PROCESSING
# ==============================================================================
openpyxl==3.1.5  # Excel export
Pillow==10.1.0   # Image processing

# ==============================================================================
# UTILITIES
# ==============================================================================
python-dateutil==2.9.0

# ==============================================================================
# PRODUCTION SERVER
# ==============================================================================
gunicorn==21.2.0
whitenoise==6.6.0  # Static file serving

# ==============================================================================
# MONITORING & LOGGING
# ==============================================================================
sentry-sdk==1.40.0
django-prometheus==2.3.1

# ==============================================================================
# LOAD TESTING (development/staging)
# ==============================================================================
locust==2.20.0

# ==============================================================================
# TESTING
# ==============================================================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
# factory-boy==3.3.0  # Optional: for test factories

# ==============================================================================
# NOTES
# ==============================================================================
# Install with: pip install -r requirements.txt
# Update all: pip install -U -r requirements.txt
# Freeze current: pip freeze > requirements-freeze.txt
